requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncio-dgram==2.1.2
attrs==22.1.0
autopep8==2.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
DateTime==4.7
Deprecated==1.2.13
discord.py==2.1.0
dnspython==2.2.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mcstatus==10.0.1
multidict==6.0.3
priority==2.0.0
psutil==5.9.4
pycodestyle==2.10.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
pymongo==4.3.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
quart==0.18.3
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
zope.interface==5.5.2
